graphql_blog/requirements.txt
fastapi==0.111.0
strawberry-graphql==0.221.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1  # Optional: for env variables (e.g., JWT secrets)
pydantic==2.7.1       # Used by FastAPI for validation
pytest==8.2.1                 # For testing GraphQL queries
httpx==0.27.0                 # FastAPI test client / GraphQL requests
